% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFindNotes.R
\name{addNote}
\alias{addNote}
\title{Add a note to an existing RData file}
\usage{
addNote(nms, note)
}
\arguments{
\item{nms}{A string that indicates the R data file created with \code{\link{digitizeRadii}} (in the current directory) to which the note should be added.}

\item{note}{The note (as a string) to be added to the R data file. If missing the user will be provided a dialog box in which to enter the note.}
}
\value{
None, but the \code{note} object in the R Data file(s) given in \code{nms} will be modified.
}
\description{
Use this to add a note to a saved R data file that was created with \code{\link{digitizeRadii}}. Can be used to flag the reading for some reason.
}
\details{
A detailed description of its use is in the "Other Features" vignette on the \href{https://fishr-core-team.github.io/RFishBC/index.html}{RFishBC website}.
}
\examples{
## See the link to the extensive documentation in the Details.

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
