% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTransect.R
\name{addTransect}
\alias{addTransect}
\title{Add a transect to an image by selecting two points.}
\usage{
addTransect(col.transect = "blue", lwd.transect = 1, lty.transect = 1)
}
\arguments{
\item{col.transect}{The color of the transect line.}

\item{lwd.transect}{The line width of the transect line.}

\item{lty.transect}{The line type of the transect line.}
}
\value{
A list of x- and y- coordinates for the two selected points.
}
\description{
Add a transect to an image by selecting a point at the focus and a point at the edge of the calcified structure. This is not required for any analyses, but it is useful for providing a landmark for identifying annuli with \code{selectPoints}.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
