% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aStandard.R
\name{aStandard}
\alias{aStandard}
\title{Finds standard intercept (a) for Fraser-Lee back-calculation model for a particular species.}
\usage{
aStandard(species)
}
\arguments{
\item{species}{A string that contains the species name for which to find teh standard intercept value.}
}
\value{
A single value from \code{\link{StdIntLit}} that is the standard intercept value (a; mm) for the species provided in \code{species}.
}
\description{
Finds standard intercept (a; mm) for Fraser-Lee back-calculation model for a particular species for all species for which a value of a has been defined.
}
\examples{
aStandard("Bluegill")
aStandard("Walleye")

}
\seealso{
\code{\link{StdIntLit}}
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
\keyword{manip}
