% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcFuns.R
\name{BCfunctions}
\alias{BCfunctions}
\alias{DahlLea}
\alias{FraserLee}
\alias{SPH}
\alias{BPH}
\alias{BiolInt}
\alias{ModFry}
\title{Specific back-calculation functions}
\usage{
DahlLea(Lcap, Ri, Rcap)

FraserLee(Lcap, Ri, Rcap, a)

SPH(Lcap, Ri, Rcap, A, B)

BPH(Lcap, Ri, Rcap, a, b)

BiolInt(Lcap, Ri, Rcap, L0p, R0p)

ModFry(Lcap, Ri, Rcap, L0p, R0p, a)
}
\arguments{
\item{Lcap}{A numeric vector of lengths-at-capture.}

\item{Ri}{A numeric vector of radial measurements to the ith annulus.}

\item{Rcap}{A numeric vector of total structure radii-at-capture.}

\item{a}{The length of the fish at the time of structure formation, the intercept of the length- on radius-at-capture regression, or, when using scales, from published “standards” for a species (Carlander 1982).}

\item{A}{The intercept of the radius- on length-at-capture regression.}

\item{B}{The slope of the radius- on length-at-capture regression.}

\item{b}{The slope of the length- on radius-at-capture regression.}

\item{L0p}{Fish length at the \dQuote{Biological Intercept}.}

\item{R0p}{Radius at the \dQuote{Biological Intercept}.}
}
\value{
A back-calculated length-at-age.
}
\description{
Provides a function to compute back-calculated lengths at previous ages for several of the most common back-calculation models: Dahl-Lea, Fraser-Lee, Body Proportional Hypothesis (BPH), Scale Proportional Hypothesis (SPH), Biological Intercept, and Modified Fry. See \code{\link{bcFuns}} for other back-calculation functions.
}
\references{
See references in \code{\link{bcFuns}}.
}
\seealso{
\code{\link{bcFuns}}.
}
