% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectRadiiData.R
\name{collectRadiiData}
\alias{collectRadiiData}
\title{Collect radial measurements from a calcified structure by interactively selecting annuli.}
\usage{
collectRadiiData(fname = file.choose(), ID, reading, suffix = reading,
  description = NULL, magnification = NULL, useScaleBar = FALSE,
  addTransect = TRUE, col.transect = "yellow", lwd.transect = 2,
  lty.transect = 1, pch.annuli = 3, col.annuli = "red",
  cex.annuli = 1.25, edgeIsAnnulus = FALSE, sepWindow = TRUE)
}
\arguments{
\item{fname}{A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).}

\item{ID}{A unique identifier for the fish or structure being examined. Will be coerced to a character.}

\item{reading}{Identifies the reading for this structure. Typically this can simply be set to \sQuote{1} if the structure will not be read multiple times. If the structure is read multiple times then this should be used to identify which time (which read) this one is. Will be coerced to a character.}

\item{suffix}{A string that will be added to the R object data file (see Details). This defaults to the same string (possibly coerced) in \code{reading}.}

\item{description}{A string that possibly contains a short description for the read of this structure.}

\item{magnification}{XXX}

\item{useScaleBar}{XXX}

\item{addTransect}{A logical that indicates whether the user will be prompted to add a linear transect to the structure image by selecting a point at the focus and a point on the margin of the structure. Adding a transect is not required, but may be useful when identifying annuli on the structure.}

\item{col.transect}{The color of the transect line if \code{addTransect=TRUE}.}

\item{lwd.transect}{The line width of the transect line if \code{addTransect=TRUE}.}

\item{lty.transect}{The line type of the transect line if \code{addTransect=TRUE}.}

\item{pch.annuli}{The plotting character of points for selected annuli.}

\item{col.annuli}{The color of points for selected annuli.}

\item{cex.annuli}{The character expansion value of points for selected annuli.}

\item{edgeIsAnnulus}{A logical that indicates whether the point at the structure margin should be considered as an annulus (\code{TRUE}) or not (\code{FALSE}). Use \code{FALSE} if the last selected point represents an incomplete year's worth of growth (i.e., \sQuote{plus-growth}).}

\item{sepWindow}{A logical that indicates whether the structure image should be opened in a new separate window (\code{=TRUE}) or not (\code{=FALSE}).}
}
\value{
A string that contains the filename (with path) of an R data object (i.e., .RData file) written to the same directory/folder as the file in \code{fname} that contains the following items in a list:
\itemize{
  \item description: The description given in \code{description}.
  \item ID: The unique identifier given in \code{ID}.
  \item image: The filename for the structure image given in \code{fname}.
  \item pts: The data.frame of x- and y-coordinates for the points (as returned directly from selecting points on the structure image).
  \item radii: A data.frame that contains the radial measurements (i.e., from the focus to the selected point) for the structure. Specifically, it contains the following items:
  \itemize{
    \item XXX
  }
}.
}
\description{
Computes radial measurements from selected points and writes all to an R data object

The user can interactively select points on a plot by clicking with the first mouse button on the image. When finished, the x- and y-coordinates for each selected point will be returned.
}
\details{
This uses \code{\link[graphics]{locator}} which will suspend activity in the console until the user indicates that they are done selecting points. The user indicates that they are done selecting points by pressing the ESCape key or right-clicking on the image and selecting Stop in Windows, pressing any mouse button other than the first (left) in X11 devices, and pressing the ESCape key in Quartz or OS X devices.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}.
}
