% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{combineData}
\alias{combineData}
\title{Combines radii data from multiple files into one data.frame}
\usage{
combineData(nms, outFormat = c("long", "wide"), deletePlusGrowth = TRUE)
}
\arguments{
\item{nms}{A character vector of R data file names created with \code{\link{digitizeRadii}}. The files must be in the current working directory (see \code{\link{getwd}} result).}

\item{outFormat}{A string that indicates the output format for the combined data. The \code{"wide"} (DEFAULT) format has one-radius-per-line (i.e., each radial measurement for a fish in on a separate row), whereas the \code{"long"} format has one-fish-per-line (i.e., each radial measurement for a fish is in a separate column).}

\item{deletePlusGrowth}{A logical that indicates whether the radial measurement that corresponds to \dQuote{plus-growth} should be deleted from the returned data.frame (\code{TRUE}; DEFAULT) or not (\code{FALSE}).}
}
\value{
A data.frame that contains the radii data created with \code{\link{digitizeRadii}} for all files given in \code{nms}.
}
\description{
Combines radial measurements made on calcified structures and saved to an R data file with \code{\link{digitizeRadii}} into a single data.frame that can then be post-processed (e.g., back-calculate length at a previous age).
}
\details{
A detailed description of its use is in \href{http://derekogle.com/RFishBC/articles/MeasureRadii/collectRadiiData.html}{this vignette} on the \href{http://derekogle.com/RFishBC/index.html}{RFishBC website}. The list of R data file names may be efficiently created with \code{\link{listFiles}} as described in that vignette. The R data file names may also be selected from a dialog box if using Windows.
}
\examples{
## See the link to the extensive documentation in the Details.

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
