% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFindNotes.R
\name{findNotes}
\alias{findNotes}
\title{Returns the notes from the chosen R Data files}
\usage{
findNotes(nms)
}
\arguments{
\item{nms}{A string (or vector of strings) that indicates the R data file(s) created with \code{\link{digitizeRadii}}. If missing the user will be provided a dialog box from which to choose the file(s). The file(s) must be in the current working directory (see \code{\link{getwd}} result). May also be a single \code{RFishBC} object created with \code{\link{digitizeRadii}}.}
}
\value{
A data.frame that contains the \code{id}, \code{reading}, and \code{notes} from the chosen R Data files.
}
\description{
Returns the notes from the chosen R data files created with \code{\link{digitizeRadii}}.
}
\details{
A detailed description of its use is in the "Other Features" vignette on the \href{http://derekogle.com/RFishBC/index.html}{RFishBC website}.
}
\examples{
## See the link to the extensive documentation in the Details.

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
