% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-internals.R
\name{.onAttach}
\alias{.onAttach}
\alias{STOP}
\alias{WARN}
\alias{CATLINE}
\alias{BULLET}
\alias{DONE}
\alias{NOTE}
\alias{iHndlFilenames}
\alias{iGetImage}
\alias{iScalingFactorFromScaleBar}
\alias{iPlaceText}
\alias{isRData}
\title{Internal functions.}
\usage{
.onAttach(lib, pkg, ...)
}
\description{
Internal functions.
}
\keyword{internal}
