% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFiles.R
\name{listFiles}
\alias{listFiles}
\title{List files with a specific extension in a folder/directory}
\usage{
listFiles(path = ".", ext = ".RData", ignore.case = TRUE, ...)
}
\arguments{
\item{path}{A string that contains the full path name for the folder/directory for which to list files. Defaults to the current working directory (see \code{\link{getwd}} result).}

\item{ext}{A string that contains the file extension pattern to match. The string should include the period.}

\item{ignore.case}{A logical for where pattern matching should be case sensitive (\code{=FALSE}) or not (\code{TRUE}; DEFAULT).}

\item{\dots}{Parameters to be given to \code{\link{list.files}}.}
}
\value{
Character vector.
}
\description{
This returns a vector with all file names with the \code{ext} extension in the \code{path} folder/directory. In \pkg{RfishBC} this is used primarily to create a list of image file names over which the user will loop or R data object names that can be given to \code{combineFiles}.
}
\details{
None yet.
}
\examples{
## None yet

}
\seealso{
\code{\link{list.files}} in base R.
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
