% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImage.R
\name{readImage}
\alias{readImage}
\title{Loads and plots any of the three bitmapped types of images.}
\usage{
readImage(fname = file.choose())
}
\arguments{
\item{fname}{A string that indicates the image to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).}
}
\value{
A string that contains the filename (with path).
}
\description{
Loads and plots any of the three bitmapped types of images as chosen by the user.
}
\details{
This uses \code{\link[readbitmap]{read.bitmap}} from the \pkg{readbitmap} package so that any of PNG, JPG, or BMP files will be automatically detected.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}. However, note that this is nearly exactly the unexported \code{ReadImg} from the \pkg{digitize} package. I did, however, add the code that produces the dialog box for choosing the file.
}
