% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviewDigitizedImages.R
\name{reviewDigitizedImages}
\alias{reviewDigitizedImages}
\title{Allows user to efficiently view images with selected points saved in multiple R data files.}
\usage{
reviewDigitizedImages(
  nms,
  deviceType,
  pch.show,
  col.show,
  cex.show,
  connect,
  col.connect,
  lwd.connect,
  col.scaleBar,
  lwd.scaleBar,
  showScaleBarLength,
  cex.scaleBar,
  showAnnuliLabels,
  annuliLabels,
  col.ann,
  cex.ann,
  offset.ann
)
}
\arguments{
\item{nms}{A string (or vector of strings) that indicates the R data file(s) created with \code{\link{digitizeRadii}}. If missing the user will be provided a dialog box from which to choose the file(s). The file(s) must be in the current working directory (see \code{\link{getwd}} result). May also be a single \code{RFishBC} object created with \code{\link{digitizeRadii}}.}

\item{deviceType}{See details in \code{\link{RFBCoptions}}.}

\item{pch.show}{See details in \code{\link{RFBCoptions}}.}

\item{col.show}{See details in \code{\link{RFBCoptions}}.}

\item{cex.show}{See details in \code{\link{RFBCoptions}}.}

\item{connect}{See details in \code{\link{RFBCoptions}}.}

\item{col.connect}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.connect}{See details in \code{\link{RFBCoptions}}.}

\item{col.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{showScaleBarLength}{See details in \code{\link{RFBCoptions}}.}

\item{cex.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{showAnnuliLabels}{See details in \code{\link{RFBCoptions}}.}

\item{annuliLabels}{See details in \code{\link{RFBCoptions}}.}

\item{col.ann}{See details in \code{\link{RFBCoptions}}.}

\item{cex.ann}{See details in \code{\link{RFBCoptions}}.}

\item{offset.ann}{See details in \code{\link{RFBCoptions}}.}
}
\value{
None.
}
\description{
Allows user to efficiently view images with points to represent annuli that were saved to multiple R data files using \code{\link{digitizeRadii}}. The user can press keyboard buttons to move between (forward and backward) images.
}
\details{
Multiple images with marked annuli recorded from \code{\link{digitizeRadii}} and saved in an RData file are viewed efficiently with this function. Appropriate RData files in the current working directory can be given as a vector (see \code{\link{listFiles}}) to \code{nms}. The image from the first first file will be displayed in a stand-along window. The image in the next file can be viewed by pressing the \sQuote{n} or \sQuote{>} keys. Previous images can be returned to by pressing the \sQuote{p} or \sQuote{<} keys. Press the \sQuote{f} or \sQuote{d} key to close the window when done reviewing the images.
}
\examples{
## None because this requires interaction from the user.
## See the link to the extensive documentation in the Details.

}
\seealso{
\code{\link{showDigitizedImage}}, \code{\link{saveDigitizedImage}}, \code{\link{digitizeRadii}}, \code{\link{RFBCoptions}}, and \code{\link{jpeg}}, \code{\link{png}}, and \code{\link{pdf}}.
}
\author{
Derek H. Ogle, \email{DerekOgle51@gmail.com}
}
