% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPoints.R
\name{selectPoints}
\alias{selectPoints}
\title{Interactively select points on an image.}
\usage{
selectPoints(pch = 3, col.pts = "red", cex.pts = 1.25)
}
\arguments{
\item{pch}{Plotting character for the selected points.}

\item{col.pts}{Color for the selected points.}

\item{cex.pts}{Character expansion factor for the selected points.}
}
\value{
A list of x- and y-coordinates for the selected points.
}
\description{
The user can interactively select points on a plot by clicking with the first mouse button on the image. When finished, the x- and y-coordinates for each selected point will be returned.
}
\details{
This uses \code{\link[graphics]{locator}} which will suspend activity in the console until the user indicates that they are done selecting points. The user indicates that they are done selecting points by pressing the ESCape key or right-clicking on the image and selecting Stop in Windows, pressing any mouse button other than the first (left) in X11 devices, and pressing the ESCape key in Quartz or OS X devices.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
