% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMBassWB1.R
\docType{data}
\name{SMBassWB1}
\alias{SMBassWB1}
\title{Fish-specific data for West Bearskin Lake Smallmouth Bass.}
\format{A data frame with 445 rows and 10 variables:
\describe{
  \item{id}{An identification number unique to each fish.}
  \item{species}{Fish species (all are \dQuote{SMB}).}
  \item{lake}{Lake (all are \dQuote{WB}).}
  \item{gear}{Capture gear (\code{E}=electrofishing, \code{T}=trapnet).}
  \item{yearcap}{Year of capture.}
  \item{lencap}{Total length-at-capture (mm)}
}}
\source{
Originally from \code{\link[FSA]{SMBassWB}} in \pkg{FSA}. Rearranged to demonstrate functions in this package.
}
\usage{
SMBassWB1
}
\description{
Fish-specific data for West Bearskin Lake Smallmouth Bass.
}
\seealso{
\code{\link{SMBassWB2}}
}
\keyword{datasets}
