% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMBassWB2.R
\docType{data}
\name{SMBassWB2}
\alias{SMBassWB2}
\title{Radial measurements for for West Bearskin Lake Smallmouth Bass.}
\format{A data frame with 181 rows and 13 variables:
\describe{
  \item{id}{An identification number unique to each fish.}
  \item{reading}{A reading specific identifier (all are \dQuote{DHO}).}
  \item{agecap}{Age-at-capture.}
  \item{radcap}{Total structure radius at capture (magnified mm).}
  \item{rad1}{Structure radius to annulus 1.}
  \item{rad2}{Structure radius to annulus 2.}
  \item{rad3}{Structure radius to annulus 3.}
  \item{rad4}{Structure radius to annulus 4.}
  \item{rad5}{Structure radius to annulus 5.}
  \item{rad6}{Structure radius to annulus 6.}
  \item{rad7}{Structure radius to annulus 7.}
  \item{rad8}{Structure radius to annulus 8.}
  \item{rad9}{Structure radius to annulus 9.}
}}
\source{
Originally from XXX in \pkg{alr4}. Rearranged to demonstrate functions in this package.
}
\usage{
SMBassWB2
}
\description{
Radial measurement data for West Bearskin Lake Smallmouth Bass. Data are in \dQuote{wide} or one-fish-per-line (with radial measurements in separate variables) format.
}
\seealso{
\code{\link{SMBassWB1}}
}
\keyword{datasets}
