% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineData.R
\name{combineData}
\alias{combineData}
\title{Combines radii data from multiple files into one data.frame}
\usage{
combineData(nms)
}
\arguments{
\item{nms}{A character vector of RData file names created with \code{\link{digitizeRadii}}. The files must be in the current working directory (see \code{\link{getwd}} result).}
}
\value{
A data.frame that contains the radii data created with \code{\link{digitizeRadii}} for all files given in \code{nms}.
}
\description{
Combines radial measurements made on calcified structures and saved to an RData file with \code{\link{digitizeRadii}} into a single data.frame that can then be post-processed (e.g., back-calculate length at a previous age).
}
\details{
A detailed description of its use is in \href{http://derekogle.com/RFishBC/articles/MeasureRadii/collectRadiiData.html}{this vignette} on the \href{http://derekogle.com/RFishBC/index.html}{RFishBC website}. The list of RData file names may be efficiently created with \code{\link{listFiles}} as described in that vignette. The RData file names may also be selected from a dialog box if using Windows.
}
\examples{
## See the link to the extensive documentation in the Details.

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
