% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialMeasurements.R
\name{combineData}
\alias{combineData}
\title{Combines radii data from multiple files into one data.frame}
\usage{
combineData(fnames, path = ".")
}
\arguments{
\item{fnames}{A character vector of file names (without the path) from which the radii data should be extracted and combined to form one synthetic data.frame. The files should all be in one folder/directory as given in \code{path}.}

\item{path}{A string that contains the full path name for the folder/directory for which to list files. Defaults to the current working directory (see \code{\link{getwd}} result).}
}
\value{
A data.frame that contains the radii data created with \code{processPoints} for all files given in \code{fnames}.
}
\description{
Combines the radii data that were created with \code{processPoints} and saved to an R object file from multiple R object files into one data.frame. This data.frame can then be post-processed.
}
\details{
None yet.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
