% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialMeasurements.R
\name{digitizeRadii}
\alias{digitizeRadii}
\title{Collect radial measurements from a calcified structure by interactively selecting annuli.}
\usage{
digitizeRadii(fname = file.choose(), id, reading, suffix, description,
  edgeIsAnnulus, sepWindow, windowSize, scaleBar, scaleBarLength, col.scaleBar,
  lwd.scaleBar, scalingFactor, addTransect, col.transect, lwd.transect, pch.sel,
  col.sel, cex.sel)
}
\arguments{
\item{fname}{A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).}

\item{id}{A unique identifier for the fish or structure being examined. Will be coerced to a character.}

\item{reading}{See details in \code{\link{RFBCoptions}}.}

\item{suffix}{See details in \code{\link{RFBCoptions}}.}

\item{description}{See details in \code{\link{RFBCoptions}}.}

\item{edgeIsAnnulus}{See details in \code{\link{RFBCoptions}}.}

\item{sepWindow}{See details in \code{\link{RFBCoptions}}.}

\item{windowSize}{See details in \code{\link{RFBCoptions}}.}

\item{scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{scaleBarLength}{See details in \code{\link{RFBCoptions}}.}

\item{col.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{scalingFactor}{See details in \code{\link{RFBCoptions}}.}

\item{addTransect}{See details in \code{\link{RFBCoptions}}.}

\item{col.transect}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.transect}{See details in \code{\link{RFBCoptions}}.}

\item{pch.sel}{See details in \code{\link{RFBCoptions}}.}

\item{col.sel}{See details in \code{\link{RFBCoptions}}.}

\item{cex.sel}{See details in \code{\link{RFBCoptions}}.}
}
\value{
A string that contains the filename (with path) of an R data object (i.e., .RData file) written to the same directory/folder as the file in \code{fname} that contains the following items in a list:
\itemize{
  \item description: The description given in \code{description}.
  \item id: The unique identifier given in \code{id}.
  \item image: The filename for the structure image given in \code{fname}.
  \item pts: The data.frame of x- and y-coordinates for the points (as returned directly from selecting points on the structure image).
  \item radii: A data.frame that contains the radial measurements (i.e., from the focus to the selected point) for the structure. Specifically, it contains the following items:
  \itemize{
    \item XXX
  }
}.
}
\description{
Computes radial measurements from selected points and writes all to an R data object

The user can interactively select points on a plot by clicking with the first mouse button on the image. When finished, the x- and y-coordinates for each selected point will be returned.
}
\details{
This uses \code{\link[graphics]{locator}} which will suspend activity in the console until the user indicates that they are done selecting points. The user indicates that they are done selecting points by pressing the ESCape key or right-clicking on the image and selecting Stop in Windows, pressing any mouse button other than the first (left) in X11 devices, and pressing the ESCape key in Quartz or OS X devices.
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}.
}
