% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findScalingFactor.R
\name{findScalingFactor}
\alias{findScalingFactor}
\title{Find scaling factor from object of known length}
\usage{
findScalingFactor(fname, knownLength, sepWindow, windowSize, col.scaleBar,
  lwd.scaleBar)
}
\arguments{
\item{fname}{A string that indicates the image (must be PNG, JPG, or BMP) to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).}

\item{knownLength}{See details in \code{\link{RFBCoptions}}.}

\item{sepWindow}{See details in \code{\link{RFBCoptions}}.}

\item{windowSize}{See details in \code{\link{RFBCoptions}}.}

\item{col.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.scaleBar}{See details in \code{\link{RFBCoptions}}.}
}
\value{
A single numeric that is the scaling factor (a multiplier that is used to convert image lengths to actual lengths).
}
\description{
This computes a scaling factor (i.e., a multiplier that is used to convert image lengths to actual lengths) from user selected endpoints of an object of known length on the image given in \code{fname} and the actual known length given in \code{knownLength}.
}
\details{
None yet.
}
\examples{
## None yet
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
