% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-internals.R
\name{.onAttach}
\alias{.onAttach}
\alias{iReadImage}
\alias{iAddTransect}
\alias{iSelectAnnuli}
\alias{iProcessAnnuli}
\title{Internal functions.}
\usage{
.onAttach(lib, pkg, ...)
}
\description{
Internal functions.
}
\keyword{internal}
