% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-internals.R
\name{STOP}
\alias{STOP}
\alias{.onAttach}
\alias{WARN}
\alias{iHndlFilename}
\alias{iHndlID}
\alias{iHndlScalingFactor}
\alias{iGetImage}
\alias{iSelectAnulli}
\alias{iFindTransect}
\alias{iShowTransect}
\alias{iSnap2Transect}
\alias{iProcessAnnuli}
\alias{iPlaceText}
\title{Internal functions.}
\usage{
STOP(..., call. = FALSE, domain = NULL)
}
\description{
Internal functions.
}
\keyword{internal}
