% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-internals.R
\name{STOP}
\alias{STOP}
\alias{.onAttach}
\alias{WARN}
\alias{iHndlFilenames}
\alias{iGetImage}
\alias{iPts2Rad}
\alias{iSnap2Transect}
\alias{iScalingFactorFromScaleBar}
\alias{iPlaceText}
\title{Internal functions.}
\usage{
STOP(..., call. = FALSE, domain = NULL)
}
\description{
Internal functions.
}
\keyword{internal}
