% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-internals.R
\name{STOP}
\alias{STOP}
\alias{.onAttach}
\alias{WARN}
\alias{iHndlfname}
\alias{iReadImage}
\alias{iProcessAnnuli}
\alias{iScaleBar}
\alias{iSelectAnulli}
\alias{iSnap2Transect}
\title{Internal functions.}
\usage{
STOP(..., call. = FALSE, domain = NULL)
}
\description{
Internal functions.
}
\keyword{internal}
