% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFishBC-package.r
\docType{package}
\name{RFishBC}
\alias{RFishBC}
\alias{RFishBC-package}
\title{RFishBC.}
\description{
A possible replacement for the fishBC package distributed by the American Fisheries Society that appears to be far out-of-date.
}
