% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDigitizedImage.R
\name{showDigitizedImage}
\alias{showDigitizedImage}
\title{Show the points that were selected at some previous time and saved to the R object file.}
\usage{
showDigitizedImage(fname, sepWindow, pch.show, col.show, cex.show, showTransect,
  col.transect, lwd.transect, col.scaleBar, lwd.scaleBar)
}
\arguments{
\item{fname}{A string that indicates the R object file to be loaded and plotted. By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (will look for this file in the current working directory unless a fully pathed name is given).}

\item{sepWindow}{See details in \code{\link{RFBCoptions}}.}

\item{pch.show}{See details in \code{\link{RFBCoptions}}.}

\item{col.show}{See details in \code{\link{RFBCoptions}}.}

\item{cex.show}{See details in \code{\link{RFBCoptions}}.}

\item{showTransect}{See details in \code{\link{RFBCoptions}}.}

\item{col.transect}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.transect}{See details in \code{\link{RFBCoptions}}.}

\item{col.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.scaleBar}{See details in \code{\link{RFBCoptions}}.}
}
\value{
None, but an image is plotted with the selected points, and possibly a putative transect, overlaid.
}
\description{
Show the points that were selected at some previous time on the structure and save to the R object file. This is useful for reexamining the selected points at a later time or overlaying selected points from multiple reads of the structure.
}
\details{
None yet
}
\examples{
## None yet

}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
