% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showDigitizedImage.R
\name{showDigitizedImage}
\alias{showDigitizedImage}
\title{Show points selected on a structure image and saved in an RData file}
\usage{
showDigitizedImage(nm, sepWindow, pch.show, col.show, cex.show, showTransect,
  col.transect, lwd.transect, col.scaleBar, lwd.scaleBar, showOrigPts = FALSE,
  pch.show2, col.show2, cex.show2)
}
\arguments{
\item{nm}{A string that indicates the RData file that contains the R object created with  created with \code{\link{digitizeRadii}} (the loaded object will contain the image and the selected points as). By default the user will be provided a dialog box from which to choose the file. Alternatively the user can supply the name of the file (file must be in the current working directory unless a fully pathed name is given).}

\item{sepWindow}{See details in \code{\link{RFBCoptions}}.}

\item{pch.show}{See details in \code{\link{RFBCoptions}}.}

\item{col.show}{See details in \code{\link{RFBCoptions}}.}

\item{cex.show}{See details in \code{\link{RFBCoptions}}.}

\item{showTransect}{See details in \code{\link{RFBCoptions}}.}

\item{col.transect}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.transect}{See details in \code{\link{RFBCoptions}}.}

\item{col.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{lwd.scaleBar}{See details in \code{\link{RFBCoptions}}.}

\item{showOrigPts}{A logical that indicates whether the original user-selected points (i.e., not snapped to the transect) should be shown or not. If the original selections were not snapped to the transect then this will be ignored.}

\item{pch.show2}{The plotting character for the original points (i.e., not snapped to the transect) selected by the user. Defaults to same as \code{pch.show}.}

\item{col.show2}{The color for the original points (i.e., not snapped to the transect) selected by the user. Defaults to same as \code{pch.show}.}

\item{cex.show2}{The character expansion value for the original points (i.e., not snapped to the transect) selected by the user. Defaults to same as \code{pch.show}.}
}
\value{
None, but an image is plotted with, at least, the selected points.
}
\description{
Show points selected on a structure image to represent annuli that were saved to an RData file using \code{\link{digitizeRadii}}. This allows the user to reexaminine the selected points or overlay selected points from multiple readings of the structure.
}
\details{
This function requires interaction from the user. A detailed description of its use is in \href{http://derekogle.com/RFishBC/articles/MeasureRadii/seeRadiiData.html}{this vignette} on the \href{http://derekogle.com/RFishBC/index.html}{RFishBC website}.
}
\examples{
## None because this requires interaction from the user.
## See the link to the extensive documentation in the Details.

}
\seealso{
\code{\link{digitizeRadii}} and \code{\link{RFBCoptions}}.
}
\author{
Derek H. Ogle, \email{derek@derekogle.com}
}
